+++ src/Sylius/Bundle/CartBundle/EventListener/CartListener.php
@@ -114,6 +114,7 @@ class CartListener implements EventSubscriberInterface
         if ($valid) {
             $this->cartManager->persist($cart);
             $this->cartManager->flush();
+
             $this->cartProvider->setCart($cart);
         }
     }
@@ -124,14 +125,5 @@ class CartListener implements EventSubscriberInterface
     private function refreshCart(CartInterface $cart)
     {
         $cart->calculateTotal();
-        $cart->setTotalItems($cart->countItems());
-        
-        // Set Total Quantity
-        $totalQuantity = 0;
-        foreach ($cart->getItems() as $item) {
-            $totalQuantity += $item->getQuantity();
-            $this->cartManager->persist($item);
-        }
-        $cart->setTotalQuantity($totalQuantity);
     }
 }