+++ src/Sylius/Bundle/CartBundle/spec/Sylius/Bundle/CartBundle/Model/CartSpec.php
@@ -11,14 +11,12 @@
 
 namespace spec\Sylius\Bundle\CartBundle\Model;
 
-use PHPSpec2\ObjectBehavior;
+use PhpSpec\ObjectBehavior;
 
 /**
- * Cart spec.
- *
  * @author Paweł Jędrzejewski <pjedrzejewski@diweb.pl>
  */
-class Cart extends ObjectBehavior
+class CartSpec extends ObjectBehavior
 {
     function it_is_initializable()
     {