+++ src/Sylius/Bundle/CartBundle/spec/Sylius/Bundle/CartBundle/Provider/CartProviderSpec.php
@@ -11,14 +11,12 @@
 
 namespace spec\Sylius\Bundle\CartBundle\Provider;
 
-use PHPSpec2\ObjectBehavior;
+use PhpSpec\ObjectBehavior;
 
 /**
- * Cart provider spec.
- *
  * @author Paweł Jędrzejewski <pjedrzejewski@diweb.pl>
  */
-class CartProvider extends ObjectBehavior
+class CartProviderSpec extends ObjectBehavior
 {
     /**
      * @param Sylius\Bundle\CartBundle\Storage\CartStorageInterface $storage